Rotation
I created a script that allows the a game object to rotate upon input. The 3 objects that are listed as children rotate accordingly, keeping their local orientation. This can be useful for switching between the stations with each station's children on them as well.
The previous storage error was solved by restarting Unity.