I WAS WRONG
In order to play the chair and in the headset, it NEEDS to be on OpenXR, and it NEEDS to be played through SteamVR.

Will try to find a workaround in the future. For now, THIS version and 2 commits prior are the working versions for the chair and headset and plugin setup.