Pulldown Levers Implemented
We have found a workaround for the levers. Through semi-hardcoding programming and using the configurable joint's SPRING component, and shrinking the Linear Limit range, the levers now pull down as far as users want and when they let go the levers bounce back up in a "spring motion" to reset in a range of it's starting position +- linear limit.
It's not the happiest workaround, but it does work well.