Levers rotate!
Levers rotate without jitters. The only issue remaining with them now is that they need to be in their ORIGINAL POSITION before rotating so that they can be used in their original range of motion each time.
(if you push a lever forward and leave it forward, then rotate the station, the lever's new range of motion will be based on it's starting position connected to the rigidbody...so the range of motion will not be the same as it used to be. Therefore, the levers MUST be reset to original position after use.)