Levers (kinda) have a better orientation after rotation.
When the station rotates, the levers are often in new orientations for their range of motion. This fix has decreased that slightly, however to ensure the lever range of motion works perfectly and remains the same after each rotation, further work will need to be done on it. But, for Beta purposes, it is sufficient.